chore(worker): rename "assets" → "tldraw_assets" for specificity
The plain "assets" name was too generic for both the SQL schema
and the worker source files — the table specifically tracks the
lifecycle of tldraw's `TLAsset` blob references (the bytes user-
uploaded images / videos resolve to via `TLAssetStore`), not
generic project assets. Rename to make the tldraw connection
explicit:

- SQL: table `assets` → `tldraw_assets`; index
  `idx_assets_document_stale_at` → `idx_tldraw_assets_document_stale_at`.
  Migration filename also renamed:
  `0002_create_assets.sql` → `0002_create_tldraw_assets.sql`.
- Worker source: `src/assets.ts` → `src/tldraw-assets.ts`,
  `src/asset-policy.ts` → `src/tldraw-asset-policy.ts`.
- Worker `package.json` exports key:
  `./asset-policy` → `./tldraw-asset-policy`.
- App imports updated:
  `anipres-worker/asset-policy` → `anipres-worker/tldraw-asset-policy`.

In-file function and variable names (e.g. `MAX_ASSET_SIZE`,
`runDocumentAssetGc`) keep the shorter `asset` form — the file
name now provides the tldraw context, so internal callers don't
need to repeat it.

Migration filename change: anyone with a local D1 that already
ran the old `0002_create_assets.sql` should wipe
`packages/worker/.wrangler/state` and re-run `wrangler dev` so
the renamed migration applies fresh. No effect on remote D1
(no production deploys yet).
