+++ packages/app/src/components/AnipresContainer.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useEffect, useRef, type ComponentProps } from "react";
 import { Anipres } from "anipres";
-import { MAX_ASSET_SIZE } from "anipres-worker/asset-policy";
+import { MAX_ASSET_SIZE } from "anipres-worker/tldraw-asset-policy";
 import type { TLStoreSnapshot } from "tldraw";
 import { useDocumentManagerContext } from "../documents/useDocumentManagerContext";
