+++ packages/worker/package.json
@@ -4,7 +4,7 @@
   "version": "0.0.1",
   "type": "module",
   "exports": {
-    "./asset-policy": "./src/asset-policy.ts"
+    "./tldraw-asset-policy": "./src/tldraw-asset-policy.ts"
   },
   "scripts": {
     "dev": "pnpm run migrate:local && wrangler dev",
