+++ packages/worker/src/tldraw-asset-policy.ts
